Robin2011_RepPackage

Documentation for Robin2011_RepPackage.

Robin2011_RepPackage.SurplusVFI

Robin2011_RepPackage.grids

Robin2011_RepPackage.homeprod

Robin2011_RepPackage.matchprod

Robin2011_RepPackage.SurplusVFI — Method

SurplusVFI(p::Matrix, z::Matrix, Π::Matrix; β::Number = 0.9466)

Performs Value Function Iteration on the Surplus Function, returns the resulting Surplus function matrix.

Takes as inputs the matrix of match productivities p, the matrix of home production z, the markov transition matrix Π, and the parameter β as a keyword argument.

β should be set to equal (1-δ)/(1+r), it's default is the value used in Robin(2011).

Robin2011_RepPackage.grids — Method

grids(; M = 500, N = 100, ρ = 0.913702234286476, σ = 0.0257,ν::Number = 2.019365636076711, μ::Number = 5.786082109731152 )

This function takes parameter values and constructs the grids and distributions of worker types and aggregate shocks, which it returns in a Dict.

Robin2011_RepPackage.homeprod — Method

homeprod(x::Vector, y::Vector; B::Number = 1, C::Number = 0.726, α::Number = 0.64, z0::Number = 0.767)

Takes the worker and aggregate state grids, as well as the parameters B, C, α, and returns the matrix of home production.

Default parameter values are those used in Robin (2011).

Robin2011_RepPackage.matchprod — Method

matchprod(x::Vector, y::Vector; B::Number = 1, C::Number = 0.725767358913686)

Takes the worker and aggregate state grids, as well as the parameters B and C, and returns the matrix of match productivities.

Default parameter values are those used in Robin (2011).