Robin2011_RepPackage

Documentation for Robin2011_RepPackage.

Robin2011_RepPackage.SurplusVFI

Robin2011_RepPackage.WageVFI

Robin2011_RepPackage.estCrit

Robin2011_RepPackage.grids

Robin2011_RepPackage.homeprod

Robin2011_RepPackage.matchprod

Robin2011_RepPackage.params_estimated

Robin2011_RepPackage.unemp_path

Robin2011_RepPackage.wage_dens_path

Robin2011_RepPackage.SurplusVFI — Method

SurplusVFI(p::Matrix, z::Matrix, Π::Matrix; β::Number = 0.9466)

Performs Value Function Iteration on the Surplus Function, returns the resulting worker x state Surplus function matrix.

Takes as inputs the matrix of match productivities p, the matrix of home production z, the markov transition matrix Π, and the parameter β as a keyword argument.

β should be set to equal (1-δ)/(1+r), it's default is the value used in Robin (2011).

Robin2011_RepPackage.WageVFI — Method

WageVFI(S::Matrix, Π::Matrix, z::Matrix; λ1 = 0.119, β::Number = 0.9466)

This function calculates equilibrium wages, as well as their associated value functions.

Wages are assigned through Bertrand Competition, resulting either in workers being paid their reservaiton wage (the monopsony wage), when being hired from unemployment, or capturing the full surplus when being poached.

The Function returns a Dict with keys:

:wmin, an i x j matrix giving the monopsony wage assigned to worker j, being hired out of unemployment in state i.

:wmax, an i x j matrix giving the full surplus wage assigned to worker j, being poached in state i.

:Wmin, an i x j x k array giving the value of worker k, being in state i, with the monopsony wage from state j.

:Wmax, an i x j x k array giving the value of worker k, being in state i, with the poacher's wage from state j.

The Function takes as arguments:

S, the surplus matrix, calculated using SurplusVFI.

Π, the markov transition matrix, calculated using grids.

z, the home production matrix, calculated using homeprod.

As well as parameters:

λ1, the probability of meeting a new employer while already employed.

β, the discounted probability of exogenenous separation, i.e. (1 - δ)/(1+r).

The default parameter values are those used in Robin(2011).

Robin2011_RepPackage.estCrit — Method

estCrit(b; M = 500, N = 100, τ = 0.5, α = 0.64, r = 0.05/4, T = 5000, burn = 1000, draw = rand(burn+T, 1), b0 = [1, 0.0226, 0.9136, 0.0579, 0.2141, 2.5296, 0.7842])

estCrit takes a set of transformed parameter values b, simulates an economy of length T+burn based on an evolution of the aggregate state given in the keyword argument draw (which should be a T+burn long series of draws from a Unif(0,1) distribution), drops the first burn periods and calculates the simulated targeted moments, before finally calculating the distance from the equivalent moments in the data, given by keyword argument b0.

b should be a vector which contains a guess of the parameters to be estimated. They should be ordered and transformed in the following way:

log(ν)

log(μ)

loginv(δ)

loginv(λ0)

atanh(ρ)

log(σ)

log(z0)

log(C)

Where logit = x -> 1/(1+exp(x)) and loginv = x -> log(1/x - 1).

External parameters are given by the following keyword arguments, with defaults set to those used in Robin (2011).

τ is the tie-breaking probability for poachers

α is the elasticity of wages to productivity - this doesn't matter for estimation but must be specified.

r is the quarterly interest rate.

Other keyword arguments include

M the length of the worker ability grid

N the length of the aggregate state grid

T the number of time periods to be used in the estimation

burn the number of additional time periods to be simulated at the beginning of the economy and then dropped to remove reliance on starting values.

b0 a vector of the moments calculated from the data.

The targeted moments are:

Average Productivity

Standard Deviation of Productivity

Autocorrelation of Productivity

Average Unemployment Rate

Standard Deviation of the Unemployment Rate

Kurtosis of the Unemployment Rate

Average Exit Rate from Unemployment

Robin2011_RepPackage.grids — Method

grids(; M = 500, N = 100, ρ = 0.913702234286476, σ = 0.0257,ν::Number = 2.019365636076711, μ::Number = 5.786082109731152 )

This function takes parameter values and constructs the grids and distributions of worker types and aggregate shocks, which it returns in a Dict, the Dict keys are:

:x, the grid of worker productivities

:y, the grid of aggregate shocks,

:Π, the Markov transition matrix,

:l, the pdf of worker types,

:F, the grid of aggregate shocks before transformed to fit the estimated distribution.

Robin2011_RepPackage.homeprod — Method

homeprod(x::Vector, y::Vector; B::Number = 1, C::Number = 0.726, α::Number = 0.64, z0::Number = 0.767)`

Takes the worker and aggregate state grids, as well as the parameters B, C, α, and returns the worker x state matrix of home production.

Default parameter values are those used in Robin (2011).

Robin2011_RepPackage.matchprod — Method

matchprod(x::Vector, y::Vector; B::Number = 1, C::Number = 0.725767358913686)

Takes the worker and aggregate state grids, as well as the parameters B and C, and returns the worker x state matrix of match productivities.

Default parameter values are those used in Robin (2011).

Robin2011_RepPackage.params_estimated — Method

params_estimated(; path = "x1.txt")

This function outputs all necessary parameters for the model, from a set of estimated parameters.

As it's input it takes a file path, which must point to a .txt file containing the parameter values correctly ordered, line separated, and transformed as in estmain.jl.

The function then performs the appropriate inverse transformations and returns the parameters as a NamedTuple.

The default file path points to those estimated by estmain.jl, stored in x1.txt. Alternatively to use those from Robin(2011), replace this with x0.txt.

The parameters must be stored in the text file in the following order with the following transformations.

log(ν)

log(μ)

loginv(δ)

loginv(λ0)

atanh(ρ)

log(σ)

log(z0)

log(C)

Where logit = x -> 1/(1+exp(x)) and loginv = x -> log(1/x - 1).

Robin2011_RepPackage.unemp_path — Method

function unemp_path(S::Matrix, statet::Vector, T::Integer; λ0::Number = 0.994544861919718, δ::Number = 0.041563759920623)

This function calculates the unemployment path for a given evolution of the aggregate state.

The Function takes as arguments:

S, the surplus matrix, calculated using SurplusVFI

statet, a given evolution of the aggregate state index through time

T, the number of periods

As well as parameters:

λ0, the rate at which an unemployed worker meets a firm

δ, the exogeneous job offer separatioin rate

The default parameter values are those used in Robin(2011).

The function returns a matrix, where each row gives a time period, and the columns give the unemployment rate for each worker type.

Robin2011_RepPackage.wage_dens_path — Method

function wage_dens_path(S::Matrix, uxt::Matrix, wd::Dict, l::Vector, U::Matrix, statet::Vector, T::Integer; λ0::Number = 0.9945, λ1::Number = 0.1193, δ::Number = 0.0416)

Wages are assigned through Bertrand Competition, resulting either in workers being paid their reservation wage (the monopsony wage), when being hired from unemployment, or capturing the full surplus when being poached. However, since workers do not receive a poaching offer every period, you may observe multiple wages at every single point in time. Furthermore, changes in the state may make it so that either the worker or the firm have a credible threat to leave/ fire leading to renegotiation of the wage.

This function calculates the wage density path for a given evolution of the aggregate state, according to the laws of motion derived in Robin (2011), giving the density of workers earning each of the possible wages at each point in time.

The Function takes as arguments:

S, the surplus matrix, calculated using SurplusVFI

uxt, the unemployment path, calculated using unemp_path

wd, a Dict including the minimum and the maximum wage in each period as well as their current value, calculated using WageVFI

l, the number of workers of each type, calculated using grids

U, the value of unemployment

statet, a given evolution of the aggregate state index through time

T, the number of periods

As well as parameters:

λ0, the rate at which an unemployed worker meets a firm

The default parameter values are those used in Robin(2011).

The function returns a four dimensional array, where each entry is the measure of workers earning that wage. The first index denotes the time period, the second the state that wage is assigned, the third the worker type that wage is assigned to, and the fourth wether the wage is the monopsony wage [:,:,:1], or the full surplus wage [:, :, :, 2].